chore: get rid of SSR for algolia 🍆
